+++ cli/src/test/kotlin/com/github/zeldigas/text2confl/cli/config/ConverterConfigTest.kt
@@ -15,7 +15,7 @@ internal class ConverterConfigTest {
         assertThat(
             ConverterConfig(
                 "Pre ", " - Post", EditorVersion.V1, null, null, null,
-                MarkdownConfiguration()
+                CodeBlockParams(), MarkdownConfiguration()
             ).titleConverter(Path(""), "test")
         )
             .isEqualTo("Pre test - Post")
@@ -26,7 +26,7 @@ internal class ConverterConfigTest {
         assertThat(
             ConverterConfig(
                 "", "", EditorVersion.V1, null, null, null,
-                MarkdownConfiguration()
+                CodeBlockParams(), MarkdownConfiguration()
             ).languageMapper.supportedLanguages
         )
             .isEqualTo(CONFLUENCE_SERVER_LANGUAGES)
@@ -37,7 +37,7 @@ internal class ConverterConfigTest {
         assertThat(
             ConverterConfig(
                 "", "", EditorVersion.V2, null, null, null,
-                MarkdownConfiguration()
+                CodeBlockParams(), MarkdownConfiguration()
             ).languageMapper.supportedLanguages
         )
             .isEqualTo(CONFLUENCE_CLOUD_LANGUAGES)